Function fatfs::format_volume

pub fn format_volume<S: ReadWriteSeek>(
    storage: &mut S, 
    options: FormatVolumeOptions
) -> Result<(), Error<S::Error>>

Create FAT filesystem on a disk or partition (format a volume)

Warning: this function overrides internal FAT filesystem structures and causes a loss of all data on provided partition. Please use it with caution. Only quick formatting is supported. To achieve a full format zero entire partition before calling this function. Supplied storage parameter cannot be seeked (internal pointer must be on position 0). To format a fragment of a disk image (e.g. partition) library user should wrap the file struct in a struct limiting access to partition bytes only e.g. fscommon::StreamSlice.

Errors

Errors that can be returned:

Error::InvalidInput will be returned if options describes an invalid file system that cannot be created. Possible reason can be requesting a fat type that is not compatible with the total number of clusters or formatting a too big storage. If sectors/clusters related options in options structure were left set to defaults this error is very unlikely to happen.

Error::Io will be returned if the provided storage object returned an I/O error.

Panics

Panics in non-optimized build if storage position returned by seek is not zero.