Struct fatfs::File

pub struct File<'a, IO: ReadWriteSeek, TP, OCC> { /* private fields */ }

A FAT filesystem file object used for reading and writing data.

This struct is created by the open_file or create_file methods on Dir.

Implementations

impl<'a, IO: ReadWriteSeek, TP, OCC> File<'a, IO, TP, OCC>

pub fn truncate(&mut self) -> Result<(), Error<IO::Error>>

Truncate file in current position.

Errors

Error::Io will be returned if the underlying storage object returned an I/O error.

Panics

Will panic if this is the root directory.

pub fn set_created(&mut self, date_time: DateTime)

👎 Deprecated

Sets date and time of creation for this file.

Note: it is set to a value from the TimeProvider when creating a file. Deprecated: if needed implement a custom TimeProvider.

pub fn set_accessed(&mut self, date: Date)

👎 Deprecated

Sets date of last access for this file.

Note: it is overwritten by a value from the TimeProvider on every file read operation. Deprecated: if needed implement a custom TimeProvider.

pub fn set_modified(&mut self, date_time: DateTime)

👎 Deprecated

Sets date and time of last modification for this file.

Note: it is overwritten by a value from the TimeProvider on every file write operation. Deprecated: if needed implement a custom TimeProvider.