Struct fatfs::FormatVolumeOptions

pub struct FormatVolumeOptions { /* fields omitted */ }

A FAT filesystem formatting options

This struct implements a builder pattern. Options are specified as an argument for format_volume function.

Implementations

impl FormatVolumeOptions

pub fn new() -> Self

Create options struct for format_volume function

Allows to overwrite many filesystem parameters. In normal use-case defaults should suffice.

pub fn bytes_per_cluster(self, bytes_per_cluster: u32) -> Self

Set size of cluster in bytes (must be dividable by sector size)

Cluster size must be a power of two and be greater or equal to sector size. If option is not specified optimal cluster size is selected based on partition size and optionally FAT type override (if specified using fat_type method).

Panics

Panics if bytes_per_cluster is not a power of two or is lower than 512.

pub fn fat_type(self, fat_type: FatType) -> Self

Set File Allocation Table type

Option allows to override File Allocation Table (FAT) entry size. It is unrecommended to set this option unless you know what you are doing. Note: FAT type is determined from total number of clusters. Changing this option can cause formatting to fail if the volume cannot be divided into proper number of clusters for selected FAT type.

pub fn bytes_per_sector(self, bytes_per_sector: u16) -> Self

Set sector size in bytes

Sector size must be a power of two and be in range 512 - 4096. Default is 512.

Panics

Panics if bytes_per_sector is not a power of two or is lower than 512.

pub fn total_sectors(self, total_sectors: u32) -> Self

Set total number of sectors

If option is not specified total number of sectors is calculated as storage device size divided by sector size.

pub fn max_root_dir_entries(self, max_root_dir_entries: u16) -> Self

Set maximal numer of entries in root directory for FAT12/FAT16 volumes

Total root directory size should be dividable by sectors size so keep it a multiple of 16 (for default sector size). Note: this limit is not used on FAT32 volumes. Default is 512.

pub fn fats(self, fats: u8) -> Self

Set number of File Allocation Tables

The only allowed values are 1 and 2. If value 2 is used the FAT is mirrored. Default is 2.

Panics

Panics if fats is outside of the range [1, 2].

pub fn media(self, media: u8) -> Self

Set media field for Bios Parameters Block

Default is 0xF8.

pub fn sectors_per_track(self, sectors_per_track: u16) -> Self

Set number of physical sectors per track for Bios Parameters Block (INT 13h CHS geometry)

Default is 0x20.

pub fn heads(self, heads: u16) -> Self

Set number of heads for Bios Parameters Block (INT 13h CHS geometry)

Default is 0x40.

pub fn drive_num(self, drive_num: u8) -> Self

Set drive number for Bios Parameters Block

Default is 0 for FAT12, 0x80 for FAT16/FAT32.

pub fn volume_id(self, volume_id: u32) -> Self

Set volume ID for Bios Parameters Block

Default is 0x12345678.

pub fn volume_label(self, volume_label: [u8; 11]) -> Self

Set volume label

Default is empty label.